requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
requests>=2.25.0
imutils>=0.5.4
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
pillow>=8.0.0
matplotlib>=3.3.0
scipy>=1.6.0
tqdm>=4.60.0
tensorboard>=2.5.0
pyyaml>=5.4.0
seaborn>=0.11.0